pyproject.toml
[tool.poetry]
name = "chain-listener"
version = "0.2.0"
description = "Universal multi-chain distributed blockchain listener SDK"
authors = ["Chain Listener Team <team@chainlistener.dev>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/chain-listener/chain-listener"
repository = "https://github.com/chain-listener/chain-listener"
documentation = "https://chain-listener.readthedocs.io"
keywords = ["blockchain", "ethereum", "listener", "events", "web3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Office/Business",
]
packages = [{include = "chain_listener", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^6.11.0"
aiohttp = "^3.9.0"
aioredis = "^2.0.0"
motor = "^3.3.0"
pydantic = "^2.5.0"
async-timeout = "^4.0.3"
pyyaml = "^6.0.1"
click = "^8.1.7"
typing-extensions = "^4.8.0"
async-limiter = "^1.0.0"
tronpy = "^0.5.0"
solana = "^0.30.2"
anchorpy = "^0.17.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.11.0"
mypy = "^1.7.1"
flake8 = "^6.1.0"
pre-commit = "^3.6.0"
isort = "^5.12.0"
bandit = "^1.7.5"
twine = "^6.2.0"
vulture = "^2.14"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"
sphinx-autodoc-typehints = "^1.24.0"

[tool.poetry.scripts]
chain-listener = "chain_listener.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["chain_listener"]
known_third_party = ["web3", "aiohttp", "aioredis", "motor", "pydantic"]

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "web3.*",
    "aiohttp.*",
    "aioredis.*",
    "motor.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=chain_listener",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=90",
]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "network: marks tests that require network access",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/chain_listener"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/virtualenv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2
fail_under = 90

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests", "build", "dist"]
skips = ["B101", "B601"]
